created preparation and deliver use case
